Added model mapper and rabbitmq container

+++ producer/src/main/java/com/mbouhda/producer/dao/AccountRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface AccountRepository extends PagingAndSortingRepository<Account, Long> {
+public interface AccountRepository extends PagingAndSortingRepository<Account, Integer> {
 
     Page<Account> findAll(Pageable pageable);
 }
